+++ src/Tool/ORM/Walker/SqlWalkerCompatForOrm3.php
@@ -32,7 +32,7 @@ use Doctrine\ORM\Query\SqlWalker;
  *
  * @internal
  */
-trait SqlWalkerCompat
+trait SqlWalkerCompatForOrm3
 {
     /**
      * Gets an executor that can be used to execute the result of this walker.